more stuff with missing nfo indicator - fixed a problem with the indicator showing up on multi-cd releases (this is now optional but turned off by default in config)

+++ zipscript/src/zipscript-c.c
@@ -100,7 +100,8 @@ void remove_nfo_indicator(char *directory) {
 		} else n++;
 	}
 	locations.nfo_incomplete = i_incomplete(incomplete_nfo_indicator, path);
-	unlink(locations.nfo_incomplete);
+	if (fileexists(locations.nfo_incomplete))
+		unlink(locations.nfo_incomplete);
 	if (k < 2) free(path[1]);
 	if (k == 0) free(path[0]);
 }
@@ -135,6 +136,11 @@ void getrelname(char *directory) {
 		sprintf(locations.link_source, "%.*s", n - 1, locations.path);
 		locations.link_target = path[0];
 		locations.incomplete = c_incomplete(incomplete_cd_indicator, path);
+#if (show_missing_nfo_in_cd == TRUE)
+		locations.nfo_incomplete = i_incomplete(incomplete_cd_nfo_indicator, path);
+#else
+		locations.nfo_incomplete = NULL;
+#endif
 		if (k < 2) free(path[1]);
 	} else {
 		raceI.misc.release_name = malloc(l[1] + 10);
